refactoring login related components
