+++ src/app/components/login-page/LoginForm/components/LoginFormAlert.tsx
@@ -1,0 +1,18 @@
+import Box from '@mui/material/Box';
+import Alert from '@mui/material/Alert';
+
+const LoginFormAlert = ({ errorMessage }: { errorMessage?: string }) => {
+  if (!errorMessage) {
+    return null;
+  }
+
+  return (
+    <Box marginBottom={2}>
+      <Alert variant="filled" icon={false} severity="error">
+        {errorMessage}
+      </Alert>
+    </Box>
+  );
+};
+
+export default LoginFormAlert;
