+++ src/app/components/login-page/LoginPageContent.tsx
@@ -1,0 +1,12 @@
+import Container from '@mui/material/Container';
+import { LoginForm } from './LoginForm';
+
+const LoginPage = () => {
+  return (
+    <Container maxWidth="sm">
+      <LoginForm />
+    </Container>
+  );
+};
+
+export default LoginPage;
