+++ src/app/components/login-page/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginPageContent } from './LoginPageContent';
